fix(tools): :bug: filter card by rarity in jp server
