+++ src/components/blocks/TeamBuilder.tsx
@@ -7,9 +7,6 @@ import {
   DialogTitle,
   DialogContent,
   FormControl,
-  InputLabel,
-  Select,
-  MenuItem,
   DialogContentText,
   DialogActions,
   Typography,
@@ -22,7 +19,6 @@ import {
   Radio,
   CircularProgress,
 } from "@mui/material";
-import makeStyles from "@mui/styles/makeStyles";
 import React, {
   useCallback,
   useEffect,
@@ -31,28 +27,17 @@ import React, {
   useState,
 } from "react";
 import { useTranslation } from "react-i18next";
-import {
-  ICardInfo,
-  IGameChara,
-  ISkillInfo,
-  ITeamCardState,
-} from "../../types.d";
+import { ICardInfo, ITeamCardState } from "../../types.d";
 import {
   useAlertSnackbar,
   useCachedData,
-  useLocalStorage,
-  useSkillMapping,
   useToggle,
   cardRarityTypeToRarity,
 } from "../../utils";
 import { CardThumb, CardThumbMedium } from "../widgets/CardThumb";
-import rarityNormal from "../../assets/rarity_star_normal.png";
-import rarityAfterTraining from "../../assets/rarity_star_afterTraining.png";
 import { teamBuildReducer } from "../../stores/reducers";
 import { useStrapi } from "../../utils/apiClient";
 import { useTeamCalc } from "../../utils/teamCalc";
-import { useCharaName } from "../../utils/i18n";
-import { attrIconMap } from "../../utils/resources";
 import { useRootStore } from "../../stores/root";
 import { observer } from "mobx-react-lite";
 import {
@@ -61,25 +46,7 @@ import {
   ISekaiCardState,
 } from "../../stores/sekai";
 import { autorun } from "mobx";
-
-const useStyle = makeStyles((theme) => ({
-  "rarity-star-img": {
-    maxWidth: "16px",
-    margin: theme.spacing(0, 0.25),
-  },
-  "dialog-paper": {
-    padding: theme.spacing(1),
-    [theme.breakpoints.up("sm")]: {
-      minWidth: "500px",
-    },
-    [theme.breakpoints.up("md")]: {
-      minWidth: "700px",
-    },
-    [theme.breakpoints.up("lg")]: {
-      minWidth: "900px",
-    },
-  },
-}));
+import FilterCardsModal from "../widgets/FilterCardsModal";
 
 const TeamBuilder: React.FC<{
   teamCards: number[];
@@ -97,9 +64,7 @@ const TeamBuilder: React.FC<{
     setTeamCardsStates,
     setTeamTotalPower,
   }) => {
-    const classes = useStyle();
     const { t } = useTranslation();
-    const getCharaName = useCharaName();
     const {
       jwtToken,
       sekai: { sekaiCardTeamMap, sekaiProfileMap, setSekaiCardTeam },
@@ -107,11 +72,8 @@ const TeamBuilder: React.FC<{
     } = useRootStore();
     const { putSekaiDecks, deleteSekaiDecks } = useStrapi(jwtToken, region);
     const { showError, showSuccess } = useAlertSnackbar();
-    const skillMapping = useSkillMapping();
 
     const [cards] = useCachedData<ICardInfo>("cards");
-    const [charas] = useCachedData<IGameChara>("gameCharacters");
-    const [skills] = useCachedData<ISkillInfo>("skills");
 
     const {
       getAreaItemBonus,
@@ -130,10 +92,6 @@ const TeamBuilder: React.FC<{
       });
     }, []);
 
-    const [characterId, setCharacterId] = useState<number>(0);
-    const [rarity, setRarity] = useState<number>(0);
-    const [attr, setAttr] = useState<string>("all");
-    const [filteredCards, setFilteredCards] = useState<ICardInfo[]>([]);
     const [saveTeamDialogVisible, setSaveTeamDialogVisible] =
       useState<boolean>(false);
     const [loadTeamDialogVisible, setLoadTeamDialogVisible] =
@@ -141,10 +99,6 @@ const TeamBuilder: React.FC<{
     const [addCardDialogVisible, setAddCardDialogVisible] =
       useState<boolean>(false);
     const [editingCard, setEditingCard] = useState<ISekaiCardState>();
-    const [isSyncCardState, setIsSyncCardState] = useLocalStorage(
-      "team-build-use-sekai-card-state",
-      false
-    );
     const [storageLocation, setStorageLocation] = useState<"local" | "cloud">(
       "local"
     );
@@ -152,7 +106,6 @@ const TeamBuilder: React.FC<{
     const [isAutoCalcBonus, toggleIsAutoCalcBonus] = useToggle(
       !!sekaiProfile?.sekaiUserProfile
     );
-    const [skillSpriteName, setSkillSpriteName] = useState("any");
 
     const [teamBuildArray, dispatchTeamBuildArray] = useReducer(
       teamBuildReducer,
@@ -179,35 +132,8 @@ const TeamBuilder: React.FC<{
     const [anchorEl, setAnchorEl] = useState<HTMLDivElement | null>(null);
     const open = useMemo(() => Boolean(anchorEl), [anchorEl]);
 
-    const filterCards = useCallback(() => {
-      if (!cards || !cards.length || !skills || !skills.length) return;
-      const skillIds = skills
-        .filter((skill) => {
-          if (skillSpriteName === "perfect_score_up")
-            return (
-              skill.skillEffects[0].activateNotesJudgmentType === "perfect"
-            );
-          else if (skillSpriteName === "life_score_up")
-            return (
-              skill.skillEffects[0].skillEffectType ===
-              "score_up_condition_life"
-            );
-          else return skill.descriptionSpriteName === skillSpriteName;
-        })
-        .map((skill) => skill.id);
-      setFilteredCards(
-        cards.filter(
-          (card) =>
-            (rarity > 0 ? card.rarity === rarity : true) &&
-            (attr !== "all" ? card.attr === attr : true) &&
-            (skillIds.length ? skillIds.includes(card.skillId) : true) &&
-            (characterId ? card.characterId === characterId : true)
-        )
-      );
-    }, [attr, cards, characterId, rarity, skillSpriteName, skills]);
-
     const handleCardThumbClick = useCallback(
-      (card: ICardInfo) => {
+      (card: ICardInfo, isSyncCardState: boolean) => {
         if (teamCards.some((tc) => tc === card.id)) {
           // showError(t("music_recommend:buildTeam.duplicatedCardError"));
           const index = teamCards.indexOf(card.id);
@@ -254,7 +180,6 @@ const TeamBuilder: React.FC<{
         setTeamCardsStates,
         showError,
         t,
-        isSyncCardState,
       ]
     );
 
@@ -732,185 +657,12 @@ const TeamBuilder: React.FC<{
             </Grid>
           )}
         </Grid>
-        <Dialog
-          open={addCardDialogVisible}
+        <FilterCardsModal
           onClose={() => setAddCardDialogVisible(false)}
-          fullWidth
-        >
-          <DialogTitle>{t("music_recommend:addCardDialog.title")}</DialogTitle>
-          <DialogContent>
-            <Grid container spacing={1} alignItems="center">
-              <Grid item>
-                <FormControl style={{ minWidth: 200 }}>
-                  <InputLabel id="add-card-dialog-select-chara-label">
-                    {t("common:character")}
-                  </InputLabel>
-                  <Select
-                    labelId="add-card-dialog-select-chara-label"
-                    value={characterId}
-                    label={t("common:character")}
-                    onChange={(e) => setCharacterId(e.target.value as number)}
-                  >
-                    <MenuItem value={0}>{t("common:all")}</MenuItem>
-                    {charas &&
-                      charas.map((chara) => (
-                        <MenuItem
-                          key={`chara-select-item-${chara.id}`}
-                          value={chara.id}
-                        >
-                          {getCharaName(chara.id)}
-                        </MenuItem>
-                      ))}
-                  </Select>
-                </FormControl>
-              </Grid>
-              <Grid item>
-                <FormControl style={{ minWidth: 120 }}>
-                  <InputLabel id="add-card-dialog-select-rarity-label">
-                    {t("common:rarity")}
-                  </InputLabel>
-                  <Select
-                    labelId="add-card-dialog-select-rarity-label"
-                    value={rarity}
-                    label={t("common:rarity")}
-                    onChange={(e) => setRarity(e.target.value as number)}
-                  >
-                    <MenuItem key={`rarity-select-item-0`} value={0}>
-                      {t("common:all")}
-                    </MenuItem>
-                    {Array.from({ length: 4 }).map((_, index) => (
-                      <MenuItem
-                        key={`rarity-select-item-${index + 1}`}
-                        value={index + 1}
-                      >
-                        {index + 1 >= 3
-                          ? Array.from({ length: index + 1 }).map((_, id) => (
-                              <img
-                                className={classes["rarity-star-img"]}
-                                src={rarityAfterTraining}
-                                alt={`star-${id}`}
-                                key={`star-${id}`}
-                              />
-                            ))
-                          : Array.from({ length: index + 1 }).map((_, id) => (
-                              <img
-                                className={classes["rarity-star-img"]}
-                                src={rarityNormal}
-                                alt={`star-${id}`}
-                                key={`star-${id}`}
-                              />
-                            ))}
-                      </MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-              </Grid>
-              <Grid item>
-                <FormControl style={{ minWidth: 60 }}>
-                  <InputLabel id="add-card-dialog-select-attr-label">
-                    {t("common:attribute")}
-                  </InputLabel>
-                  <Select
-                    labelId="add-card-dialog-select-attr-label"
-                    value={attr}
-                    label={t("common:attribute")}
-                    onChange={(e) => setAttr(e.target.value as string)}
-                  >
-                    <MenuItem key={`attr-select-item-0`} value={"all"}>
-                      {t("common:all")}
-                    </MenuItem>
-                    {["cute", "mysterious", "cool", "happy", "pure"].map(
-                      (name) => (
-                        <MenuItem key={`attr-select-item-${name}`} value={name}>
-                          <img
-                            className={classes["rarity-star-img"]}
-                            src={attrIconMap[name as "cool"]}
-                            alt={`attr-${name}`}
-                            key={`attr-${name}`}
-                          />
-                        </MenuItem>
-                      )
-                    )}
-                  </Select>
-                </FormControl>
-              </Grid>
-              <Grid item>
-                <FormControl style={{ minWidth: 200 }}>
-                  <InputLabel id="add-card-dialog-select-skill-label">
-                    {t("card:skillName")}
-                  </InputLabel>
-                  <Select
-                    labelId="add-card-dialog-select-skill-label"
-                    value={skillSpriteName}
-                    label={t("card:skillName")}
-                    onChange={(e) =>
-                      setSkillSpriteName(e.target.value as string)
-                    }
-                  >
-                    <MenuItem value={"any"}>{t("common:all")}</MenuItem>
-                    {skillMapping.map(
-                      ({ name, descriptionSpriteName }, index) => (
-                        <MenuItem
-                          key={`skill-select-item-${index + 1}`}
-                          value={descriptionSpriteName}
-                        >
-                          {name}
-                        </MenuItem>
-                      )
-                    )}
-                  </Select>
-                </FormControl>
-              </Grid>
-              <Grid item>
-                <Button
-                  variant="contained"
-                  color="primary"
-                  onClick={() => filterCards()}
-                >
-                  {t("common:search")}
-                </Button>
-              </Grid>
-            </Grid>
-            {!!filteredCards.length && (
-              <Grid container spacing={1}>
-                <Grid item>
-                  <FormControlLabel
-                    control={
-                      <Switch
-                        checked={isSyncCardState}
-                        onChange={(_, checked) => setIsSyncCardState(checked)}
-                      />
-                    }
-                    label={t("team_build:use_sekai_card_state") as string}
-                    disabled={
-                      !sekaiCardTeam ||
-                      !sekaiCardTeam.cards ||
-                      !sekaiCardTeam.cards.length
-                    }
-                  />
-                </Grid>
-              </Grid>
-            )}
-            <Grid container direction="row" spacing={1}>
-              {filteredCards.map((card) => (
-                <Grid
-                  key={`filtered-card-${card.id}`}
-                  item
-                  xs={4}
-                  md={3}
-                  lg={2}
-                >
-                  <CardThumb
-                    cardId={card.id}
-                    onClick={() => handleCardThumbClick(card)}
-                    style={{ cursor: "pointer" }}
-                  />
-                </Grid>
-              ))}
-            </Grid>
-          </DialogContent>
-          {/* <DialogActions></DialogActions> */}
-        </Dialog>
+          sekaiCardTeam={sekaiCardTeam}
+          onCardSelected={handleCardThumbClick}
+          open={addCardDialogVisible}
+        />
         <Dialog
           open={saveTeamDialogVisible}
           onClose={() => setSaveTeamDialogVisible(false)}
